Refactor: move StatusCache to state package and improve state helpers
