+++ state/keyboard.py
@@ -8,6 +8,17 @@ from __future__ import annotations
 from datetime import datetime
 from typing import Any
 
+KEY_MAP = {
+    "__u__": "menu_unmapped",
+    "__l__": "menu_lan_local",
+    "__o__": "menu_open_ports",
+    "__t__": "menu_cache_terminal",
+    "__c__": "menu_clear_cache",
+    "__r__": "menu_recheck_geoip",
+    "__h__": "menu_help",
+    "__a__": "menu_about",
+    "__esc__": "escape",
+}
 
 def build_key_action(value: str) -> dict[str, Any] | None:
     """Build key action payload from capture value."""
@@ -15,19 +26,7 @@ def build_key_action(value: str) -> dict[str, Any] | None:
         return None
 
     token = value.split("|", 1)[0]
-    key_map = {
-        "__u__": "menu_unmapped",
-        "__l__": "menu_lan_local",
-        "__o__": "menu_open_ports",
-        "__t__": "menu_cache_terminal",
-        "__c__": "menu_clear_cache",
-        "__r__": "menu_recheck_geoip",
-        "__h__": "menu_help",
-        "__a__": "menu_about",
-        "__esc__": "escape",
-    }
-
-    action = key_map.get(token)
+    action = KEY_MAP.get(token)
     if not action:
         return None
